Updated event handling to use the object type of the data structure rather than the json received for resolving conversions.
Updated event types from int to long.

Added error handling for ExitStarted event and tested it.

Also tested and fixed the following events:

All Config Events
All Scene Events
All Input Events
All Scene Item Events

+++ COBWEBS-Client/Events/CurrentEvenTransitionDurationChangedEventArgs.cs
@@ -8,6 +8,6 @@ namespace COBWEBS_Client.Events
 {
 	public class CurrentEvenTransitionDurationChangedEventArgs : EventArgs
 	{
-		public int transitionDuration { get; set; }
+		public long transitionDuration { get; set; }
 	}
 }
